fix(events): declare explicit consumer group
By default the Redis messaging implementation will use the topic name as
the group name.
If we have multiple consumers listening to the same
topic for different purposes without specifying a group name, they will all be considered in the same consumer group. Thus messages are not distributed to all consumers.

Signed-off-by: Fatih Acar <fatih@opsmill.com>